create clinker
