+++ pkgs/native_toolchain_c/lib/src/cbuilder/run_cbuilder.dart
@@ -132,18 +132,6 @@ class RunCBuilder {
     }
   }
 
-  Future<void> runLinker() async {
-    assert(linkerOptions != null);
-    final toolInstance_ = await linker();
-    final tool = toolInstance_.tool;
-    if (tool.isClangLike || tool.isLdLike) {
-      await linkClangLike(tool: toolInstance_);
-      return;
-    } else {
-      throw UnimplementedError('This package does not know how to run $tool.');
-    }
-  }
-
   Future<void> runClangLike({required ToolInstance tool}) async {
     // Clang for Windows requires the MSVC Developer Environment.
     final environment = await _resolver.resolveEnvironment(tool);
@@ -218,180 +206,6 @@ class RunCBuilder {
         environment,
       );
     }
-  }
-
-  Future<void> linkClangLike({required ToolInstance tool}) async {
-    // Clang for Windows requires the MSVC Developer Environment.
-    final environment = await _resolver.resolveEnvironment(tool);
-
-    assert(staticLibrary == null);
-
-    final targetIosSdk =
-        codeConfig.targetOS == OS.iOS ? codeConfig.iOS.targetSdk : null;
-
-    // The Android Gradle plugin does not honor API level 19 and 20 when
-    // invoking clang. Mimic that behavior here.
-    // See https://github.com/dart-lang/native/issues/171.
-    final int? targetAndroidNdkApi;
-    if (codeConfig.targetOS == OS.android) {
-      final minimumApi =
-          codeConfig.targetArchitecture == Architecture.riscv64 ? 35 : 21;
-      targetAndroidNdkApi = max(codeConfig.android.targetNdkApi, minimumApi);
-    } else {
-      targetAndroidNdkApi = null;
-    }
-
-    final targetIOSVersion =
-        codeConfig.targetOS == OS.iOS ? codeConfig.iOS.targetVersion : null;
-    final targetMacOSVersion =
-        codeConfig.targetOS == OS.macOS ? codeConfig.macOS.targetVersion : null;
-
-    final Iterable<String> sourceFiles =
-        sources.map((e) => e.toFilePath()).toList();
-    final outFile =
-        dynamicLibrary != null ? outDir.resolveUri(dynamicLibrary!) : null;
-
-    await runProcess(
-      executable: tool.uri,
-      environment: environment,
-      arguments: [
-        if (installName != null) ...[
-          '-install_name',
-          installName!.toFilePath(),
-        ],
-        if (targetMacOSVersion != null) ...[
-          '-macos_version_min',
-          '$targetMacOSVersion',
-        ],
-        '-dylib',
-        ...['-arch', appleLdArchs[codeConfig.targetArchitecture]!],
-        if (pic != null)
-          if (tool.tool.isClangLike && codeConfig.targetOS != OS.windows) ...[
-            if (pic!) ...[
-              if (dynamicLibrary != null) '-fPIC',
-              if (executable != null) ...[
-                // Generate position-independent code for executables.
-                '-fPIE',
-                // Tell the linker to generate a position-independent
-                // executable.
-                '-pie',
-              ],
-            ] else ...[
-              // Disable generation of any kind of position-independent code.
-              '-fno-PIC',
-              '-fno-PIE',
-              // Tell the linker to generate a position-dependent executable.
-              if (executable != null) '-no-pie',
-            ],
-          ] else if (tool.tool.isLdLike) ...[
-            if (pic!) ...[
-              if (executable != null) '--pie',
-            ] else ...[
-              if (executable != null) '--no-pie',
-            ],
-          ],
-        if (std != null) '-std=$std',
-        if (language == Language.cpp) ...[
-          '-x',
-          'c++',
-          '-l',
-          cppLinkStdLib ?? defaultCppLinkStdLib[codeConfig.targetOS]!,
-        ],
-        if (codeConfig.targetOS == OS.macOS) ...[
-          '-syslibroot',
-          (await macosSdk(logger: logger)).toFilePath(),
-        ],
-        if (optimizationLevel != OptimizationLevel.unspecified)
-          optimizationLevel.clangFlag(),
-        ...linkerOptions?.toLinkerSyntax(
-              tool.tool,
-              (sourceFiles.any((source) => source.endsWith('.a')) ||
-                          linkerOptions!.includeAllSymbols) &&
-                      OS.current == OS.linux
-                  ? ['--whole-archive']
-                  : [],
-            ) ??
-            [],
-        // Support Android 15 page size by default, can be overridden by
-        // passing [flags].
-        if (codeConfig.targetOS == OS.android) '-Wl,-z,max-page-size=16384',
-        ...flags,
-        for (final MapEntry(key: name, :value) in defines.entries)
-          if (value == null) '-D$name' else '-D$name=$value',
-        for (final include in includes) '-I${include.toFilePath()}',
-        for (final forcedInclude in forcedIncludes)
-          '-include${forcedInclude.toFilePath()}',
-        ...(linkerOptions != null
-            ? switch (OS.current) {
-              OS.linux => sourceFiles,
-              OS.macOS => sourceFiles.expand(
-                (source) =>
-                    source.endsWith('.a')
-                        ? [
-                          if (linkerOptions!.includeAllSymbols)
-                            linkerOptions!.toLinkerSyntax(tool.tool, [
-                              '-force_load',
-                            ]).first,
-                          source,
-                        ]
-                        : [source],
-              ),
-              OS() => sourceFiles,
-            }
-            : sourceFiles),
-        if (language == Language.objectiveC) ...[
-          for (final framework in frameworks) ...['-framework', framework],
-        ],
-        if (executable != null) ...[
-          '-o',
-          outDir.resolveUri(executable!).toFilePath(),
-        ] else if (dynamicLibrary != null) ...[
-          '-o',
-          outFile!.toFilePath(),
-        ] else if (staticLibrary != null) ...[
-          '-c',
-          '-o',
-          outFile!.toFilePath(),
-        ],
-        ...linkerOptions?.toLinkerSyntax(tool.tool, [
-              ...[
-                ...linkerOptions!.linkerFlags,
-                if (OS.current == OS.macOS) '-lSystem',
-              ],
-              if (linkerOptions!.stripDebug)
-                switch (OS.current) {
-                  OS.linux => '--strip-debug',
-                  OS.macOS => '-S',
-                  OS() => throw UnimplementedError(),
-                },
-              if (linkerOptions!.gcSections)
-                OS.current == OS.macOS ? '-dead_strip' : '--gc-sections',
-              if (linkerOptions!.linkerScript != null && OS.current == OS.linux)
-                '--version-script=${linkerOptions!.linkerScript!.toFilePath()}',
-              if ((sourceFiles.any((source) => source.endsWith('.a')) ||
-                      linkerOptions!.includeAllSymbols) &&
-                  OS.current == OS.linux)
-                '--no-whole-archive',
-            ]) ??
-            [],
-        if (executable != null || dynamicLibrary != null) ...[
-          if (codeConfig.targetOS case OS.android || OS.linux)
-            // During bundling code assets are all placed in the same directory.
-            // Setting this rpath allows the binary to find other code assets
-            // it is linked against.
-            if (linkerOptions != null)
-              '-rpath=\$ORIGIN'
-            else
-              '-Wl,-rpath=\$ORIGIN',
-          for (final directory in libraryDirectories)
-            '-L${directory.toFilePath()}',
-          for (final library in libraries) '-l$library',
-        ],
-      ],
-      logger: logger,
-      captureOutput: false,
-      throwOnUnexpectedExitCode: true,
-    );
   }
 
   /// [toolInstance] is either a compiler or a linker.
